Modify persone.rb to person.rb file
